Moving extension class to separate file
